# backend/openapi.yml
openapi: 3.0.0
info:
  title: Flask SQLAlchemy RESTful API
  version: 1.0.0
  description: |-
    產生的 OpenAPI 文件，對應 Flask + SQLAlchemy 所有 RESTful API。
    
    - 所有 API 回傳 JSON。
servers:
  - url: http://127.0.0.1:8000
    description: Development server
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
...
paths:
...
  /minio/ensure_bucket:
    post:
      tags: [MinIO]
      summary: 確保 MinIO bucket 存在
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bucket:
                  type: string
                  example: rsrp
      responses:
        '200':
          description: Bucket 存在或已建立
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '404':
          description: Bucket 不存在
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Bucket 不存在"
  /minio/upload_json:
    post:
      tags: [MinIO]
      summary: 上傳 JSON 字串到 MinIO
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bucket:
                  type: string
                  example: rsrp
                object_name:
                  type: string
                  example: test.json
                json_str:
                  type: string
                  example: '{"a":1,"b":2}'
      responses:
        '200':
          description: 結果
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /minio/get_json:
    post:
      tags: [MinIO]
      summary: 取得 MinIO 內的 JSON 字串
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bucket:
                  type: string
                  example: rsrp
                object_name:
                  type: string
                  example: test.json
      responses:
        '200':
          description: 結果
          content:
            application/json:
              schema:
                type: object
                properties:
                  json:
                    type: string
                    example: '{"a":1,"b":2}'
  /minio/delete_object:
    post:
      tags: [MinIO]
      summary: 刪除 MinIO 內的物件
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bucket:
                  type: string
                  example: rsrp
                object_name:
                  type: string
                  example: test.json
      responses:
        '200':
          description: 刪除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          description: 缺少參數
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Missing parameter
        '404':
          description: 找不到 bucket 或 object
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
        '500':
          description: 刪除失敗
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
  
  # AODT APIs
...